requirements.txt
pyyaml>=6.0         
pandas>=1.5.0        
numpy>=1.23.0       
matplotlib>=3.6.0   
seaborn>=0.12.0      
plotly>=5.11.0       
scipy>=1.9.0
scikit-optimize>=0.9.0 
traci>=1.15.0
sumolib>=1.15.0
stable-baselines3>=2.0.0
gymnasium>=0.28.0
sb3-contrib>=2.0.0
tqdm>=4.65.0
scipy>=1.10.0
scikit-learn>=1.3.0
streamlit>=1.38